+++ returnn/tf/engine.py
@@ -1879,15 +1879,6 @@ class Engine(EngineBase):
         self.updater.set_learning_rate(eval_learning_rate, session=self.tf_session)
         return True
 
-    def _do_save(self):
-        """
-        :return: whether to perform save on disk in this process. e.g. for Horovod rank != 0, do not save.
-        :rtype: bool
-        """
-        import returnn.util.basic
-
-        return returnn.util.basic.should_write_to_disk(config=self.config)
-
     def _is_dataset_evaluated(self, name):
         """
         Check via self.learning_rate_control.
@@ -2131,118 +2122,6 @@ class Engine(EngineBase):
                     print("Last epoch model not yet evaluated on dev. Doing that now.", file=log.v4)
                     self.eval_model(skip_already_evaluated=True)
                     break
-
-    def cleanup_old_models(self, ask_for_confirmation=False):
-        """
-        :param bool ask_for_confirmation: if True, will ask the user interactively to confirm
-        """
-        if not self._do_save():
-            return
-        from returnn.util.basic import CollectionReadCheckCovered, human_bytes_size, confirm
-        from itertools import count
-
-        opts = CollectionReadCheckCovered(self.config.get_of_type("cleanup_old_models", dict, {}))
-        existing_models = self.get_existing_models(config=self.config)
-        if hasattr(self, "learning_rate_control"):
-            lr_control = self.learning_rate_control
-        else:
-            lr_control = load_learning_rate_control_from_config(self.config)
-        epochs = sorted(existing_models.keys())
-        if not epochs:
-            print("Cannot cleanup models, no models found.", file=log.v2)
-            return
-        keep_last_n = opts.get("keep_last_n", 2)
-        keep_best_n = opts.get("keep_best_n", 4)
-        assert keep_last_n >= 1 and keep_best_n >= 0
-        if max(keep_last_n, keep_best_n) >= len(epochs):
-            print(
-                (
-                    "Only %i epochs stored so far and keeping last %i epochs and best %i epochs,"
-                    " thus not cleaning up any epochs yet."
-                )
-                % (len(epochs), keep_last_n, keep_best_n),
-                file=log.v2,
-            )
-            return
-        keep_epochs = set()  # type: typing.Set[int]
-        default_keep_pattern = set()
-        if epochs[-1] <= 10:
-            keep_every = 4
-            keep_doubles_of = 5
-        elif epochs[-1] <= 50:
-            keep_every = 20
-            keep_doubles_of = 5
-        elif epochs[-1] <= 100:
-            keep_every = 40
-            keep_doubles_of = 10
-        else:
-            keep_every = 80
-            keep_doubles_of = 20
-        for i in count(1):
-            n = keep_every * i
-            if n > epochs[-1]:
-                break
-            default_keep_pattern.add(n)
-        for i in count():
-            n = keep_doubles_of * (2**i)
-            if n > epochs[-1]:
-                break
-            default_keep_pattern.add(n)
-        keep_epochs.update(opts.get("keep", default_keep_pattern))
-        keep_epochs.update(epochs[-keep_last_n:])
-        score_keys = set()  # e.g. "dev_error", "dev_score", etc.
-        # Collect all possible score keys. Note that we could have different ones for different epochs.
-        for data in lr_control.epoch_data.values():
-            score_keys.update(data.error.keys())
-        assert score_keys
-        score_keys = sorted(score_keys)
-        score_values = {key: [] for key in score_keys}
-        for epoch in epochs:
-            epoch_scores = lr_control.epoch_data[epoch].error
-            for key in epoch_scores.keys():
-                score_values[key].append(epoch_scores[key])
-        for key in list(score_keys):
-            scores = score_values[key]
-            if min(scores) == max(scores):
-                print(
-                    "Ignoring score key %r because all epochs have the same value %r." % (key, scores[0]), file=log.v3
-                )
-                score_keys.remove(key)
-                score_values.pop(key)
-        # Actually, terminology is a bit confusing. We call it "score" here (and elsewhere), but it's a loss,
-        # so the maximum value is the worst possible value.
-        worst_score_values = {key: max(scores) for (key, scores) in score_values.items()}
-        for key in score_keys:
-            scores = sorted(
-                [(lr_control.epoch_data[epoch].error.get(key, worst_score_values[key]), epoch) for epoch in epochs]
-            )
-            scores = scores[:keep_best_n]
-            keep_epochs.update([v[1] for v in scores])
-        keep_epochs.intersection_update(epochs)
-        if len(keep_epochs) == len(epochs):
-            print("%i epochs stored so far and keeping all." % len(epochs), file=log.v2)
-            return
-        remove_epochs = sorted(set(epochs).difference(keep_epochs))
-        assert remove_epochs
-        if len(epochs) > 6:
-            epoch_summary = "[%s, ..., %s]" % (", ".join(map(str, epochs[:3])), ", ".join(map(str, epochs[-3:])))
-        else:
-            epoch_summary = str(epochs)
-        print(
-            "We have stored models for epochs %s and keep epochs %s." % (epoch_summary, sorted(keep_epochs)),
-            file=log.v3,
-        )
-        print("We will delete the models of epochs %s." % (remove_epochs,), file=log.v3)
-        opts.assert_all_read()
-        if self.config.bool("dry_run", False):
-            print("Dry-run, will not delete models.", file=log.v2)
-            return
-        if ask_for_confirmation:
-            confirm("Delete those models?", exit_on_false=True)
-        count_bytes = 0
-        for epoch in remove_epochs:
-            count_bytes += self.delete_model(existing_models[epoch])
-        print("Deleted %s." % human_bytes_size(count_bytes), file=log.v2)
 
     def check_uninitialized_vars(self):
         """
